Optionally use PDC to map tags to releases.
In Fedora, when presented with a koji tag, our code who just do some string
coercion to figure out what release it was associated with.  We would guess.

The internal instance of PDC however has a field that lets you look up a
release given a tag.  This patch lets us use that field if configured, but
still retains the guesswork code for when PDC cannot provide that info.